Expose AnonymousObserver
In certain scenarios, it might be useful to be able to create the anonymous observer outside the Subscribe overloads.

For example, the official guidance on DiagnosticSource filtering recommends creating an anonymous observer to combine with the predicate to filter events.

There may be other scenarios too where not having to manually create an IObserver<T> could be useful.

Fixes #8