Match System.Reactive namespaces
This makes RxFree a 100% compatible drop-in replacement if only the provided types are used in a given project. Likewise, if the need arises for feature not provided in RxFree, it's trivial to "migrate" to System.Reactive: just uninstall RxFree and install System.Reactive and start using whatever is missing from this project.

Fixes #9